[#15] improving third test with dpckan_test module
